pyproject.toml
# =====================================================================
# For a full TOML configuration guide, check the Flower docs:
# https://flower.ai/docs/framework/how-to-configure-pyproject-toml.html
# =====================================================================
[build-system]
requires = ["hatchling", "setuptools>=42", "wheel"]
build-backend = "hatchling.build"

[project]
name = "g-pfl-ids"
version = "1.0.0"
description = "Graph-Driven Personalized Federated Learning for Unsupervised Intrusion Detection in Non-IID IoT Systems"
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"

# # Include package data
# [project.include-packages]
# include = ["configs/*"]

# Dependencies for your Flower App
dependencies = [
    "flwr>=1.8.0",
    "flwr[simulation]>=1.21.0",
    "flwr-datasets",
    "torch>=2.2.0.1",
    "torchvision>=0.17.0.1",
    "torchaudio>=2.2.0.1",

    # core libs
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.2",
    "scipy",
    "networkx>=3.0",
    "tqdm",
    "pyyaml",
    "matplotlib",
    "seaborn",
    "tqdm>=4.62.0",

    # PyTorch Geometric and its dependencies
    "torch_geometric>=2.4.0",
    "pyg_lib",
    "torch_scatter",
    "torch_sparse",
    "torch_cluster",
    "torch_spline_conv",

    # Optional (baselines, plotting)
    "xgboost",
    "joblib",
    "python-dotenv",
    "scikit-plot",
    "wandb",  # optional for logging
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "isort>=5.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "Oladele D.A."
module = "src.client_app"

# Point to your ServerApp and ClientApp objects
# Format: "<module>:<object>"
[tool.flwr.app.components]
serverapp = "src.server_app:app"
clientapp = "src.client_app:app"

# Custom config values accessible via `context.run_config`
[tool.flwr.app.config]
num-server-rounds = 20
fraction-fit = 0.5
fraction-evaluate = 1.0
min-fit-clients = 2
min-evaluate-clients = 2
min-available-clients = 10
local-epochs = 30
mode = "oneclass"
model_type = "hybrid" # Options for ONECLASS: {"deepsvdd", "gae", "hybrid"}
strategy = "fedavg"  # Options: fedavg, fedprox
learning-rate = 0.001

# Default federation to use when running the app
[tool.flwr.federations]
default = "local-simulation"

# Local simulation federation with 10 virtual SuperNodes
[tool.flwr.federations.local-simulation]
options.num-supernodes = 10

# Remote federation example for use with SuperLink
[tool.flwr.federations.remote-federation]
address = "<SUPERLINK-ADDRESS>:<PORT>"
insecure = true  # Remove this line to enable TLS
# root-certificates = "<PATH/TO/ca.crt>"  # For TLS setup
